+++ tests/test_sensors.py
@@ -25,17 +25,22 @@ def clear_sensors_file_fixture():
     yield
     os.remove(sensors.SENSOR_FILE)
 
+# define mock_sensor_data
+mock_sensor_data = [
+    {"serviceName": "Thermostat", "uniqueId": "thermostat-uid", "type": "Thermostat", "accessoryInformation": {"Serial Number": "somfythermostat:193910495402#1"}},
+    {"serviceName": "LightSensor", "uniqueId": "light-uid", "type": "LightSensor", "accessoryInformation": {"Serial Number": "io:8548271"}},
+    {"serviceName": "TemperatureSensor", "uniqueId": "temp-uid", "type": "TemperatureSensor", "accessoryInformation": {"Serial Number": "io:9229676"}},
+    {"serviceName": "TemperatureSenso2r", "uniqueId": "temp2-uid", "type": "TemperatureSensor", "accessoryInformation": {"Serial Number": "io:5041775"}},
+    {"serviceName": "HumiditySensor", "uniqueId": "humidity-uid", "type": "HumiditySensor", "accessoryInformation": {"Serial Number": "somfythermostat:193910495402#3"}}
+]
+
 
 def test_filter_and_organize_sensor_data(mock_env_vars, tmp_path):
-    with patch.object(HomebridgeClient, 'accessories', return_value=[
-        {"serviceName": "Thermostat", "uniqueId": "thermostat-uid", "type": "Thermostat"},
-        {"serviceName": "LightSensor", "uniqueId": "light-uid", "type": "LightSensor"},
-        {"serviceName": "TemperatureSensor", "uniqueId": "temp-uid", "type": "TemperatureSensor"},
-        {"serviceName": "HumiditySensor", "uniqueId": "humidity-uid", "type": "HumiditySensor"}
-    ]):
+    with patch.object(HomebridgeClient, 'accessories', return_value=mock_sensor_data):
         data = sensors.data()
         assert data == {
             "outdoor": "temp-uid",
+            "outdoor2": "temp2-uid",
             "sunlight": "light-uid",
             "thermostat": "thermostat-uid",
             "humidity": "humidity-uid"
@@ -43,12 +48,7 @@ def test_filter_and_organize_sensor_data(mock_env_vars, tmp_path):
 
 
 def test_write_sensor_data_to_file(mock_env_vars, tmp_path):
-    with patch.object(HomebridgeClient, 'accessories', return_value=[
-        {"serviceName": "Thermostat", "uniqueId": "thermostat-uid", "type": "Thermostat"},
-        {"serviceName": "LightSensor", "uniqueId": "light-uid", "type": "LightSensor"},
-        {"serviceName": "TemperatureSensor", "uniqueId": "temp-uid", "type": "TemperatureSensor"},
-        {"serviceName": "HumiditySensor", "uniqueId": "humidity-uid", "type": "HumiditySensor"}
-    ]):
+    with patch.object(HomebridgeClient, 'accessories', return_value=mock_sensor_data):
         data = sensors.data()
         sensor_file_path = tmp_path / "sensors.json"
         with open(sensor_file_path, "w") as sensor_file: